Split voting test module into model and property
