+++ ouroboros-consensus/ouroboros-consensus.cabal
@@ -616,6 +616,8 @@ test-suite consensus-test
     Test.Consensus.MiniProtocol.ObjectDiffusion.PerasCert.Smoke
     Test.Consensus.MiniProtocol.ObjectDiffusion.Smoke
     Test.Consensus.Peras.Voting
+    Test.Consensus.Peras.Voting.Model
+    Test.Consensus.Peras.Voting.Smoke
     Test.Consensus.Peras.WeightSnapshot
     Test.Consensus.Util.MonadSTM.NormalForm
     Test.Consensus.Util.Pred
